Refactor UI components and enhance functionality
This commit introduces several updates across various components, with the goal of improving the UI/UX.

Key changes include:
- **Header Component**: Added navigation links with icons, improved search functionality, and mobile responsiveness. I am personally not yet happy with the look of it currently yet, as it feels cramped and dated to me.
- **Advanced Search Component**: Enhanced input handling for both text and select elements, ensuring a more intuitive user experience.
- **Prompt Card Components**: Introduced TooltipProvider for reaction buttons to replace the hover logic, improved accessibility with aria-labels, and refined styles for better visual hierarchy. Replaced the native buttons with the Button component across the PromptCard sub-components. Incoming updates to this component, as it is currently relies on a messy presentation methodology that is unfit for mobile, and uncomfortable on desktop.
- **Filter Bar**: Replaced select elements with a more consistent UI component for category and sorting options.
- **Loading State**: Updated to include skeleton loading states that mimic the structure of prompt cards alright enough for now, updates for it soon, now that I see the prompts visually.
- **Not Found Page**: Replaced Links with Buttons, made the '404' non-interactable, and fit the 404 text inside its background wrapper to avoid the awkward look I personally thought it had.
- **Waitlist Page**: Integrated header, as all other pages have it, and there is no method of navigation otherwise.
- Some other more minor changes that should help out with consistent usability and aesthetics.

The goal was to improve the overall usability, intuitiveness, and aesthetics of the site; while ensuring that all components are cohesive and responsive. On the developer end, the code should be now cleaner and more consistent across components and pages, more consistently making use of UI components with less native implementations.

I recommend testing and validation to confirm that all functionalities work as intended and that no regressions have been introduced, as I was partly flying blind here and might not have all the experience required to make guarantees.

There are comments included in this commit that will not serve a purpose in the long-run, there are existing comments that are unnecessary, and a solely comments-focused run-through of the codebase is recommended eventually to clean up the code.